Start rating component; refactor to accommodate shared type reorgnization
